Building the main model
